AddressFuncs

Description

A collection of tested on and off chain plutus address functions.

Documentation

checkValidMultisig :: TxInfo -> [PubKeyHash] -> Integer -> Bool

A simple multisig validation. Using the script context information, validate every public key hash signature. If the public key hash is a tx signer then increment the counter. At the end check if the number of signers is greater than or equal to some threshold integer. This assumes the list of public key hashes are known at the time of validation.

checkValidMultisig scriptContextInfo listOfPkhs threshold

Testing: Test.Groups.Address

isAddrGettingPaidExactly :: [TxOut] -> Address -> Value -> Bool

Search a list of TxOut for TxOut with a specific address and value. This is a simple way to check if there exist an output utxo that has exactly some value and is being sent to a known address.

Testing: Test.Groups.Address

isAddrHoldingToken :: [TxOut] -> Address -> CurrencySymbol -> TokenName -> Integer -> Bool

Search a list of TxOut for a TxOut with a specific address that is hodling an exact amount of of a singular token. This is a great function when only a token is known but not the minimum amount of ada that is travelling with that token during a transaction.

Testing: Test.Groups.Address

createAddress :: PubKeyHash -> PubKeyHash -> Address

Create a proper Address type. A transaction can input only the public key hash of some wallet. This applies for payment and staking. Inside plutus an address is a combination of both pkhs and is not bech32. The function accounts for both types of addresses. Not to be used with validator hashes.

>>> createAddress "acab" ""
Address {addressCredential = PubKeyCredential acab, addressStakingCredential = Nothing}
>>> createAddress "acab" "beef"
Address {addressCredential = PubKeyCredential acab, addressStakingCredential = Just (StakingHash (PubKeyCredential beef))}

Testing: Test.Groups.Address

isNInputs :: [TxInInfo] -> Integer -> Bool

Count the number of inputs that have datums of any kind. A script input is spendable if and only if it has some kind of datum, either embedded or inline. Thus limiting the number of script inputs is simply limiting the number of inputs that have some type of datum.

Testing: Test.Groups.Address

isNOutputs :: [TxOut] -> Integer -> Bool

Count the number of outputs that have datums of any kind. This is great at limiting the number of outputs in a transaction that can have datums. An example is looking an outputs coming back to the script.

Testing: Test.Groups.Address