CryptoFuncs

Description

A collection of tested on and off chain plutus crypto functions.

Documentation

verifyDiscretLogarithm :: Integer -> Integer -> Integer -> Integer -> Integer -> Integer -> Bool

Given a generator g of prime order q and agreed upon r and c constants, verify that g^z = g^r * u^c mod q. This should allow a user to select some secret, x, and place it into the form u = g^x for plublic viewing. A user can then pass in z = r + c*x, to prove that x is known without revealing x.

Testing: Test.Groups.Crypto

merkleTree :: [BuiltinByteString] -> BuiltinByteString

Calculate the merkle tree from a list of bytestrings.

If the list of strings is empty then the Merkle tree will return the hash of the empty bytestring. If the length is odd then the empty bytestring is appended to the list such that it can be split into pairs. When the list has an even length, the first step is hashing each element of the original list. Then in a pairwise fashion the merkle tree is built, ultimately reutnring the hash of the final object.

Testing: Test.Groups.Crypto