ListFuncs

Description

A collection of tested on and off chain plutus list functions.

Documentation

addTwoLists :: [Integer] -> [Integer] -> [Integer]

Add pairs of elements together from two lists into a new list. The length of the list will be the shortest input list. Empty list inputs return empty list outputs.

>>> addTwoList [1,2,3] [3,2,1]
[4,4,4]

Testing: Test.Groups.List

subTwoLists :: [Integer] -> [Integer] -> [Integer]

Subtracts pairs of elements together from two lists into a new list. The length of the list will be the shortest input list.

Only positive integers.

>>> subTwoLists [3,2,1] [1,2,3]
[2,0,0]

Testing: Test.Groups.List

multiplyTwoLists :: [Integer] -> [Integer] -> [Integer]

Multiply pairs of elements together from two lists into a new list. The length of the list will be the shortest input list. Empty list inputs return empty list outputs.

>>> multiplyTwoList [1,2,3] [1,2,3]
[1,4,9]

Testing: Test.Groups.List

multiplyAList :: [Integer] -> Integer -> [Integer]

Multiply each element of a list by some scaler into a new list. The resulting list is the same length as the input list. Empty list inputs return empty list outputs.

>>> multiplyAList [1,2,3] 3
[3,6,9]

Testing: Test.Groups.List

divideTwoLists :: [Integer] -> [Integer] -> [Integer]

Divide pairs of elements together from two lists into a new list. The length of the list will be the shortest input list. Empty list inputs return empty list outputs. The first list is divded by the second list.

>>> divideTwoLists [4,9,14] [1,3,7]
[4,3,2]

Testing: Test.Groups.List

divideAList :: [Integer] -> Integer -> [Integer]

Divide each element of a list by a scaler into a new list. The length of the list will be equal to the input list. Empty list inputs return empty list outputs. Dividing by zero changes nothing.

>>> divideAList [3,6,9] 3
[1,2,3]

Testing: Test.Groups.List

replicate :: [Integer] -> Integer -> [Integer]

Replicates a list l, n times. Simple replicate function.

>>> replicate [1,2,3] 2
[1,2,3,1,2,3]

Testing: Test.Groups.List