MathFuncs

Description

A collection of tested on and off chain plutus math functions.

Documentation

pow :: Integer -> Integer -> Integer

Calculates x to the power of n using the exponentiation by squaring method.

>>> pow 513 3
135005697

Testing: Test.Groups.Math

powmod :: Integer -> Integer -> Integer -> Integer

Calculates the x to the power of n modulo q.

>>> powmod 1425 2434 71
57

Testing: Test.Groups.Math

percentage :: Integer -> Integer -> Integer

Calculate the percentage of some integer. Integer division applies here so there are only so many possible outcomes. If the pct is zero then the amount is zero.

>>> percentage 1234567890 40
30864197

Testing: Test.Groups.Math

baseQ :: Integer -> Integer -> [Integer]

Write an integer n in base q and return it as a list of integers. This is the general algorithm for base conversion for any arbitrary base and number.

>>> baseQ 42 3
[1,1,2,0]

Testing: Test.Groups.Math

logOfXInBaseB :: Integer -> Integer -> Integer

The log of x in base b for plutus. This is the integer division version.

>>> logOfXInBaseB 42 3
3

Testing: Test.Groups.Math