StringFuncs

Description

A collection of tested on and off chain plutus string functions.

Documentation

byteStringAsIntegerList :: BuiltinByteString -> [Integer]

Converts a hex encoded string into a list of integers for hardcoding. Without type validation there is not way to hardcode bytestring unless they are represented as a list of integers that can be mapped back into a bytestring on-chain.

>>> pkh = "31ec74a9f86884e7a16f8fc30840f7f409c08b91e93d2be3a4377442"
>>> byteStringAsIntegerList pkh
[49,236,116,169,248,104,132,231,161,111,143,195,8,64,247,244,9,192,139,145,233,61,43,227,164,55,116,66]

Testing: Test.Groups.String

integerAsByteString :: Integer -> BuiltinByteString

Convert an integer into a string. This converts an integer into a list of integers representing the digit in base 10. This list is feed into a mapping that converts the integer into a bytestring. Great for creating incremental token names.

>>> "HelloWorld_" <> integerAsByteString 42
"HelloWorld_42"

Testing: Test.Groups.String

hash :: BuiltinByteString -> BuiltinByteString

Compute the sha3_256 hash of some bytestring.

>>> hash "Hello, World!"
"\SUB\241zfN?\168\228\EM\184\186\ENQ\194\161s\SYN\157\247ab\165\162\134\224\196\ENQ\180`\212x\247\239"
>>> DatumHash $ hash "Hello, World!"
1af17a664e3fa8e419b8ba05c2a173169df76162a5a286e0c405b460d478f7ef

Testing: Test.Groups.String

createBuiltinByteString :: [Integer] -> BuiltinByteString

Creates a proper BuiltinByteString type from a list of integers. This is a hack solution to by pass the loss of type validation. It should be used with byteStringAsIntegerList. This allows a bytestring to be hard coded into a contract at compile time.

createBuiltinByteString $ byteStringAsIntegerList pkh

Testing: Test.Groups.String

convertByteStringToInteger :: BuiltinByteString -> Integer

Take in a bytestring and converts it to a number via a product. Similar to creating an integer list but instead takes the product of all the values. The product is limited to numbers less than 2^64 - 1.

Testing: Test.Groups.String