TimeFuncs

Description

A collection of tested on and off chain plutus time functions.

Documentation

isTxInsideInterval :: Interval POSIXTime -> POSIXTimeRange -> Bool

Check if the validity range is inside a time interval. The validity range must be completely contained within the time range. That is, a contains b if for every entry s, if member s b then member s a. This is designed to be used with the script context inside a validation script.

txValidityRange :: POSIXTimeRange
txValidityRange = txInfoValidRange $ scriptContextTxInfo scriptContext

Testing: Test.Groups.Time

isTxOutsideInterval :: Interval POSIXTime -> POSIXTimeRange -> Bool

Check if the validity range of the tx is outside of the time interval, that is, whether there is a value that is not a member of both intervals. This does not allow for validity ranges to be on the boundary of the time range. This is designed to be used with the script context inside a validation script.

txValidityRange :: POSIXTimeRange
txValidityRange = txInfoValidRange $ scriptContextTxInfo scriptContext

Testing: Test.Groups.Time

lockUntilTimeInterval :: Integer -> Interval POSIXTime

Pick the locking interval, assume negative inf to endingTime. This should be used to lock some utxo for all time up until some point in the future. This is designed to be used with the script context inside a validation script.

>>> lockUntilTimeInterval 42
Interval {ivFrom = LowerBound NegInf True, ivTo = UpperBound (Finite (POSIXTime {getPOSIXTime = 42})) True}

Testing: Test.Groups.Time

lockBetweenTimeInterval :: Integer -> Integer -> Interval POSIXTime

The time interval for the tx to be locked. This is designed to create a time interval between point a and point b.

>>> lockBetweenTimeInterval 17 19
Interval {ivFrom = LowerBound (Finite (POSIXTime {getPOSIXTime = 17})) True, ivTo = UpperBound (Finite (POSIXTime {getPOSIXTime = 19})) True}

Testing: Test.Groups.Time