UsefulFuncs

Description

A collection of tested on-chain and off-chain plutus functions.

Documentation

integerAsByteString :: Integer -> BuiltinByteString

Convert an integer into a string. This converts an integer into a list of integers representing the digit in base 10. This list is feed into a mapping that converts the integer into a bytestring. Great for creating incremental token names.

>>> "HelloWorld_" <> integerAsByteString 42
"HelloWorld_42"

Testing: Test.Groups.String

byteStringAsIntegerList :: BuiltinByteString -> [Integer]

Converts a hex encoded string into a list of integers for hardcoding. Without type validation there is not way to hardcode bytestring unless they are represented as a list of integers that can be mapped back into a bytestring on-chain.

>>> pkh = "31ec74a9f86884e7a16f8fc30840f7f409c08b91e93d2be3a4377442"
>>> byteStringAsIntegerList pkh
[49,236,116,169,248,104,132,231,161,111,143,195,8,64,247,244,9,192,139,145,233,61,43,227,164,55,116,66]

Testing: Test.Groups.String

createBuiltinByteString :: [Integer] -> BuiltinByteString

Creates a proper BuiltinByteString type from a list of integers. This is a hack solution to by pass the loss of type validation. It should be used with byteStringAsIntegerList. This allows a bytestring to be hard coded into a contract at compile time.

createBuiltinByteString $ byteStringAsIntegerList pkh

Testing: Test.Groups.String

hash :: BuiltinByteString -> BuiltinByteString

Compute the sha3_256 hash of some bytestring.

>>> hash "Hello, World!"
"\SUB\241zfN?\168\228\EM\184\186\ENQ\194\161s\SYN\157\247ab\165\162\134\224\196\ENQ\180`\212x\247\239"
>>> DatumHash $ hash "Hello, World!"
1af17a664e3fa8e419b8ba05c2a173169df76162a5a286e0c405b460d478f7ef

Testing: Test.Groups.String

convertByteStringToInteger :: BuiltinByteString -> Integer

Take in a bytestring and converts it to a number via a product. Similar to creating an integer list but instead takes the product of all the values. The product is limited to numbers less than 2^64 - 1.

Testing: Test.Groups.String

convertToString :: [Integer] -> BuiltinByteString -> BuiltinByteString

Converts a list of integers into a string by mapping integers to letters up to the value of 64. This is unlike createBuiltinByteString which uses ascii codes to create a string where this is a one to one mapping.

>>> convertToString [0..64] ""
"0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ+/"

Testing: Test.Groups.String

isAddrGettingPaidExactly :: [TxOut] -> Address -> Value -> Bool

Search a list of TxOut for TxOut with a specific address and value. This is a simple way to check if there exist an output utxo that has exactly some value and is being sent to a known address.

Testing: Test.Groups.Address

isAddrHoldingExactlyToken :: [TxOut] -> Address -> CurrencySymbol -> TokenName -> Integer -> Bool

Search a list of TxOut for a TxOut with a specific address that is hodling an exact amount of of a singular token. This is a great function when only a token is known but not the minimum amount of ada that is travelling with that token during a transaction.

Testing: Test.Groups.Address

isAddrHoldingAtLeastToken :: [TxOut] -> Address -> CurrencySymbol -> TokenName -> Integer -> Bool

Search a list of TxOut for a TxOut with a specific address that is hodling at least the amount of of a singular token. This is a great function when only a token is known but not the minimum amount of ada that is travelling with that token during a transaction.

Testing: Test.Groups.Address

isNInputs :: [TxInInfo] -> Integer -> Bool

Count the number of inputs that have datums of any kind. A script input is spendable if and only if it has some kind of datum, either embedded or inline. Thus limiting the number of script inputs is simply limiting the number of inputs that have some type of datum.

Testing: Test.Groups.Address

isNOutputs :: [TxOut] -> Integer -> Bool

Count the number of outputs that have datums of any kind. This is great at limiting the number of outputs in a transaction that can have datums. An example is looking an outputs coming back to the script.

Testing: Test.Groups.Address

createAddress :: PubKeyHash -> PubKeyHash -> Address

Create a proper Address type. A transaction can input only the public key hash of some wallet. This applies for payment and staking. Inside plutus an address is a combination of both pkhs and is not bech32. The function accounts for both types of addresses. Not to be used with validator hashes.

>>> createAddress "acab" ""
Address {addressCredential = PubKeyCredential acab, addressStakingCredential = Nothing}
>>> createAddress "acab" "beef"
Address {addressCredential = PubKeyCredential acab, addressStakingCredential = Just (StakingHash (PubKeyCredential beef))}

Testing: Test.Groups.Address

checkValidMultisig :: TxInfo -> [PubKeyHash] -> Integer -> Bool

A simple multisig validation. Using the script context information, validate every public key hash signature. If the public key hash is a tx signer then increment the counter. At the end check if the number of signers is greater than or equal to some threshold integer. This assumes the list of public key hashes are known at the time of validation.

checkValidMultisig scriptContextInfo listOfPkhs threshold

Testing: Test.Groups.Address

lockUntilTimeInterval :: Integer -> Interval POSIXTime

Pick the locking interval, assume negative inf to endingTime. This should be used to lock some utxo for all time up until some point in the future. This is designed to be used with the script context inside a validation script.

>>> lockUntilTimeInterval 42
Interval {ivFrom = LowerBound NegInf True, ivTo = UpperBound (Finite (POSIXTime {getPOSIXTime = 42})) True}

Testing: Test.Groups.Time

lockBetweenTimeInterval :: Integer -> Integer -> Interval POSIXTime

The time interval for the tx to be locked. This is designed to create a time interval between point a and point b.

>>> lockBetweenTimeInterval 17 19
Interval {ivFrom = LowerBound (Finite (POSIXTime {getPOSIXTime = 17})) True, ivTo = UpperBound (Finite (POSIXTime {getPOSIXTime = 19})) True}

Testing: Test.Groups.Time

isTxOutsideInterval :: Interval POSIXTime -> POSIXTimeRange -> Bool

Check if the validity range of the tx is outside of the time interval, that is, whether there is a value that is not a member of both intervals. This does not allow for validity ranges to be on the boundary of the time range. This is designed to be used with the script context inside a validation script.

txValidityRange :: POSIXTimeRange
txValidityRange = txInfoValidRange $ scriptContextTxInfo scriptContext

Testing: Test.Groups.Time

isTxInsideInterval :: Interval POSIXTime -> POSIXTimeRange -> Bool

Check if the validity range is inside a time interval. The validity range must be completely contained within the time range. That is, a contains b if for every entry s, if member s b then member s a. This is designed to be used with the script context inside a validation script.

txValidityRange :: POSIXTimeRange
txValidityRange = txInfoValidRange $ scriptContextTxInfo scriptContext

Testing: Test.Groups.Time

pow :: Integer -> Integer -> Integer

Calculates x to the power of n using the exponentiation by squaring method.

>>> pow 513 3
135005697

Testing: Test.Groups.Math

powmod :: Integer -> Integer -> Integer -> Integer

Calculates the x to the power of n modulo q.

>>> powmod 1425 2434 71
57

Testing: Test.Groups.Math

logOfXInBaseB :: Integer -> Integer -> Integer

The log of x in base b for plutus. This is the integer division version.

>>> logOfXInBaseB 42 3
3

Testing: Test.Groups.Math

baseQ :: Integer -> Integer -> [Integer]

Write an integer n in base q and return it as a list of integers. This is the general algorithm for base conversion for any arbitrary base and number.

>>> baseQ 42 3
[1,1,2,0]

Testing: Test.Groups.Math

percentage :: Integer -> Integer -> Integer

Calculate the percentage of some integer. Integer division applies here so there are only so many possible outcomes. If the pct is zero then the amount is zero.

>>> percentage 1234567890 40
30864197

Testing: Test.Groups.Math

replicate :: [Integer] -> Integer -> [Integer]

Replicates a list l, n times. Simple replicate function.

>>> replicate [1,2,3] 2
[1,2,3,1,2,3]

Testing: Test.Groups.List

addTwoLists :: [Integer] -> [Integer] -> [Integer]

Add pairs of elements together from two lists into a new list. The length of the list will be the shortest input list. Empty list inputs return empty list outputs.

>>> addTwoList [1,2,3] [3,2,1]
[4,4,4]

Testing: Test.Groups.List

subTwoLists :: [Integer] -> [Integer] -> [Integer]

Subtracts pairs of elements together from two lists into a new list. The length of the list will be the shortest input list.

Only positive integers.

>>> subTwoLists [3,2,1] [1,2,3]
[2,0,0]

Testing: Test.Groups.List

multiplyTwoLists :: [Integer] -> [Integer] -> [Integer]

Multiply pairs of elements together from two lists into a new list. The length of the list will be the shortest input list. Empty list inputs return empty list outputs.

>>> multiplyTwoList [1,2,3] [1,2,3]
[1,4,9]

Testing: Test.Groups.List

multiplyAList :: [Integer] -> Integer -> [Integer]

Multiply each element of a list by some scaler into a new list. The resulting list is the same length as the input list. Empty list inputs return empty list outputs.

>>> multiplyAList [1,2,3] 3
[3,6,9]

Testing: Test.Groups.List

divideTwoLists :: [Integer] -> [Integer] -> [Integer]

Divide pairs of elements together from two lists into a new list. The length of the list will be the shortest input list. Empty list inputs return empty list outputs. The first list is divded by the second list.

>>> divideTwoLists [4,9,14] [1,3,7]
[4,3,2]

Testing: Test.Groups.List

divideAList :: [Integer] -> Integer -> [Integer]

Divide each element of a list by a scaler into a new list. The length of the list will be equal to the input list. Empty list inputs return empty list outputs. Dividing by zero changes nothing.

>>> divideAList [3,6,9] 3
[1,2,3]

Testing: Test.Groups.List

checkForCurrencySymbol :: [CurrencySymbol] -> CurrencySymbol -> Bool

Check if the policy id is in the list of policy id from some value. If nothing is found or if the input is an empty list then its false.

This is designed to be combined with Value.symbols. It's a great way to check if a specific policy id is contained within some validating value.

-- The list of CurrencySymbols of a Value.
symbols :: Value -> [CurrencySymbol]

Testing: Test.Groups.Value

isNRedeemers :: [Redeemer] -> Integer

Count how many redeemers are being used inside the tx. This can be used to force the number of redeemers used when spending multiple script utxos. This is designed to be used with the script context inside a validation script.

toList :: Foldable t => t a -> [a]

-- A Map is Foldable.
txInfoRedeemers :: Map ScriptPurpose Redeemer

redeemers :: [Redeemer]
redeemers = toList $ txInfoRedeemers $ scriptContextTxInfo scriptContext

Testing: Test.Groups.Value

adaValue :: Integer -> Value

Create a pure ada singleton with a known amount of lovelace. The amount may be positive or negative depending on the use case.

>>> adaValue 123
Value (Map [(,Map [("",123)])])

Testing: Test.Groups.Value

verifyDiscretLogarithm :: Integer -> Integer -> Integer -> Integer -> Integer -> Integer -> Bool

Given a generator g of prime order q and agreed upon r and c constants, verify that g^z = g^r * u^c mod q. This should allow a user to select some secret, x, and place it into the form u = g^x for plublic viewing. A user can then pass in z = r + c*x, to prove that x is known without revealing x.

Testing: Test.Groups.Crypto

merkleTree :: [BuiltinByteString] -> BuiltinByteString

Calculate the merkle tree from a list of bytestrings.

If the list of strings is empty then the Merkle tree will return the hash of the empty bytestring. If the length is odd then the empty bytestring is appended to the list such that it can be split into pairs. When the list has an even length, the first step is hashing each element of the original list. Then in a pairwise fashion the merkle tree is built, ultimately reutnring the hash of the final object.

Testing: Test.Groups.Crypto