ValueFuncs

Description

A collection of tested on and off chain plutus value functions.

Documentation

adaValue :: Integer -> Value

Create a pure ada singleton with a known amount of lovelace. The amount may be positive or negative depending on the use case.

>>> adaValue 123
Value (Map [(,Map [("",123)])])

Testing: Test.Groups.Value

checkForCurrencySymbol :: [CurrencySymbol] -> CurrencySymbol -> Bool

Check if the policy id is in the list of policy id from some value. If nothing is found or if the input is an empty list then its false.

This is designed to be combined with Value.symbols. It's a great way to check if a specific policy id is contained within some validating value.

-- The list of CurrencySymbols of a Value.
symbols :: Value -> [CurrencySymbol]

Testing: Test.Groups.Value

isNRedeemers :: [Redeemer] -> Integer

Count how many redeemers are being used inside the tx. This can be used to force the number of redeemers used when spending multiple script utxos. This is designed to be used with the script context inside a validation script.

toList :: Foldable t => t a -> [a]

-- A Map is Foldable.
txInfoRedeemers :: Map ScriptPurpose Redeemer

redeemers :: [Redeemer]
redeemers = toList $ txInfoRedeemers $ scriptContextTxInfo scriptContext

Testing: Test.Groups.Value