UsefulFuncs

Documentation

integerAsByteString :: Integer -> BuiltinByteString

Convert an integer into a string.

Testing: Test.Groups.Helpers

byteStringAsIntegerList :: BuiltinByteString -> [Integer]

"Converts a hex encoded string into a list of integers for hardcoding."

Testing: Test.Groups.Helpers

isAddrGettingPaidExactly :: [TxOut] -> Address -> Value -> Bool

Search each TxOut for an addr and value.

Testing: Test.Groups.Address

isAddrHoldingToken :: [TxOut] -> Address -> CurrencySymbol -> TokenName -> Integer -> Bool

Search each TxOut for an addr and value.

Testing: Test.Groups.Address

isNInputs :: [TxInInfo] -> Integer -> Bool

Count the number of inputs that have datums of any kind.

Testing: Test.Groups.Address

isNOutputs :: [TxOut] -> Integer -> Bool

Count the number of outputs that have datums of any kind.

Testing: Test.Groups.Address

isNRedeemers :: [Redeemer] -> Integer

Count how many redeemers are being used inside the tx.

Testing: Test.Groups.Value

createAddress :: PubKeyHash -> PubKeyHash -> Address

Create a proper Address type.

Testing: Test.Groups.Address

createBuiltinByteString :: [Integer] -> BuiltinByteString

Creates a proper BuiltinByteString type.

Testing: Test.Groups.Helpers

lockUntilTimeInterval :: Integer -> Interval POSIXTime

Pick the locking interval, assume negative inf to endingTime.

Testing: Test.Groups.Time

lockBetweenTimeInterval :: Integer -> Integer -> Interval POSIXTime

The time interval for the tx to be locked.

Testing: Test.Groups.Time

isTxOutsideInterval :: Interval POSIXTime -> POSIXTimeRange -> Bool

Check if the validity range of the tx is outside of the time interval

Testing: Test.Groups.Time

isTxInsideInterval :: Interval POSIXTime -> POSIXTimeRange -> Bool

Check if the validity range is inside the time interval

Testing: Test.Groups.Time

hash :: BuiltinByteString -> BuiltinByteString

Compute the sha3 256 hash of some byte string

Testing: Test.Groups.Helpers

checkForCurrencySymbol :: [CurrencySymbol] -> CurrencySymbol -> Bool

Check if the policy id is in the list of policy id from some value

Testing: Test.Groups.Value

logOfXInBaseB :: Integer -> Integer -> Integer

The log of x in base b in plutus

Testing: Test.Groups.Helpers

replicate :: [Integer] -> Integer -> [Integer]

Replicates a list, l, n times.hash

Testing: Test.Groups.Helpers

checkValidMultisig :: TxInfo -> [PubKeyHash] -> Integer -> Bool

Simple Multisig, using the info check all the pkh sigs

Testing: Test.Groups.Address

baseQ :: Integer -> Integer -> [Integer]

Write an integer in base Q and return a list of integers.

Testing: Test.Groups.Helpers

addTwoLists :: [Integer] -> [Integer] -> [Integer]

add pairs of elements together from two lists into a new list

Testing: Test.Groups.List

subTwoLists :: [Integer] -> [Integer] -> [Integer]

subtracts pairs of elements together from two lists into a new list

Testing: Test.Groups.List

multiplyTwoLists :: [Integer] -> [Integer] -> [Integer]

multiply pairs of elements together from two lists into a new list

Testing: Test.Groups.List

multiplyAList :: [Integer] -> Integer -> [Integer]

multiply each element of a list by some scaler into a new list

Testing: Test.Groups.List

convertByteStringToInteger :: BuiltinByteString -> Integer

Take in a bytestring and convert it to a number by the product of hex number

Testing: Test.Groups.Helpers

divideTwoLists :: [Integer] -> [Integer] -> [Integer]

divide pairs of elements together from two lists into a new list

Testing: Test.Groups.List

divideAList :: [Integer] -> Integer -> [Integer]

divide each element of a list by some scaler into a new list

Testing: Test.Groups.List